Give the remaining loose Subsystems/ files and Engine/'s input files homes too
Follow-up to the previous commit after feedback that leaving
EC_CameraSystem/EC_SpatialSystem/EC_TransformSystem/ControlSystem at
Subsystems root (my original call, reasoning each had no closer
relative) still left things untidy, and that Engine/'s own root
(Controller/Keyboard/Mouse/MouseButton alongside Config/Timer/
EC_Engine) needed the same treatment.

- Subsystems/{Camera,Control,Spatial,Transform}/: one folder per
  system, matching the CollisionSystems/Scripting precedent rather
  than inventing thematic groupings (e.g. bundling Camera+Control)
  that weren't asked for.
- Engine/Controllers/: Controller.*, Keyboard.*, Mouse.*, MouseButton.h -
  the input-device family, grouped together since Controller.h is
  already the shared base Keyboard/Mouse both derive from.

Note: Engine/Subsystems/Spatial/ and the pre-existing top-level
Spatial/ (EC_Frustum, EC_SpatialGrid, RayQueryHit - frustum culling/
spatial partitioning) are different concerns that happen to share a
name at different nesting depths; flagging in case that's confusing
rather than silently leaving it.
