+++ src/Engine/EC_Engine.cpp
@@ -1,7 +1,7 @@
 #include "EC_Engine.h"
-#include "Engine/Subsystems/EC_SpatialSystem.h"
-#include "Engine/Subsystems/EC_TransformSystem.h"
-#include "Engine/Subsystems/EC_CameraSystem.h"
+#include "Engine/Subsystems/Spatial/EC_SpatialSystem.h"
+#include "Engine/Subsystems/Transform/EC_TransformSystem.h"
+#include "Engine/Subsystems/Camera/EC_CameraSystem.h"
 #include "Engine/Subsystems/Scripting/EC_LuaScriptingSystem.h"
 #include "Engine/Subsystems/CollisionSystems/EC_CollisionSystem.h"
 #include "Engine/Subsystems/CollisionSystems/EC_PhysicsSystem.h"
