+++ src/Engine/Subsystems/Camera/EC_CameraSystem.h
@@ -1,10 +1,10 @@
 #pragma once
-#include "EC_System.h"
+#include "Engine/Subsystems/EC_System.h"
 
-class EC_SpatialSystem : public EC_System {
+class EC_CameraSystem : public EC_System {
 public:
-    EC_SpatialSystem();
-    virtual ~EC_SpatialSystem();
+    EC_CameraSystem();
+    virtual ~EC_CameraSystem();
 
     virtual void init(ECXMessenger& messenger, EC_Game& game) override;
     virtual void update(const float& deltaTimeS, EC_Game& game) override;
