+++ src/Engine/Subsystems/Spatial/EC_SpatialSystem.cpp
@@ -1,4 +1,4 @@
-#include "Engine/Subsystems/EC_SpatialSystem.h"
+#include "Engine/Subsystems/Spatial/EC_SpatialSystem.h"
 #include "Entity/EC_DOD_EntityManager.h"
 #include "Components/EC_DOD_Components.h"
 #include <glm/gtc/matrix_transform.hpp>
