+++ src/Messaging/MouseEvent.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Engine/MouseButton.h"
+#include "Engine/Controllers/MouseButton.h"
 
 class MouseEvent
 {
